sukebei/helpers/requirements.txt
nhentai-api
requests
